Update OneDrive removal script to exclude WinSxS
Also, be more specific about what actually does get deleted. This fix is because of an issue where if "OneDrive.svg" gets deleted, Windows 11 will fail the installation of the monthly CU, and the log shows it's because the .svg file is missing.
This change now only attempts to remove OneDrive folder, OneDrive.ico, and OneDrive.exe. The fix for this has already been tested and confirmed by another project.